codebook_agent/pyproject.toml
[tool.poetry]
name = "codebook_agent"
version = "0.0.1"
description = "Agent for analyzing municipal codebooks using LangGraph."
authors = ["Kush Bhuwalka, Evan Brooks, Logan McLaughlin"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
langchain = "^0.3.21"
python = ">=3.9,<4.0"
langgraph = ">=0.2.6"
python-dotenv = ">=1.0.1"
requests = ">=2.31.0"
langchain-openai = ">=0.0.5"
langchain-core = ">=0.1.27"
setuptools = "^78.1.0"
beautifulsoup4 = "^4.13.3"
tenacity = "^9.1.2"
more-itertools = "^10.6.0"
qdrant-client = "^1.13.3"
langchain-qdrant = "^0.2.0"
nest-asyncio = "^1.6.0"
lxml = "^5.3.2"
openevals = "^0.0.18"

[tool.poetry.group.dev.dependencies]
mypy = ">=1.11.1"
ruff = ">=0.6.1"

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"